// packages/core/src/generated/schema.ts
export const ROOTS: Record<string, RootDescriptor> = {
  GeneralErrorHeaderResponse: {
    name: 'GeneralErrorHeaderResponse',
    ns: 'common',
    type: 'common:GeneralErrorHeaderResponse',
  },
  GeneralErrorResponse: {
    name: 'GeneralErrorResponse',
    ns: 'api',
    type: 'api:GeneralErrorResponse',
  },
  GeneralExceptionResponse: {
    name: 'GeneralExceptionResponse',
    ns: 'common',
    type: 'common:GeneralExceptionResponse',
  },
  InvoiceAnnulment: { name: 'InvoiceAnnulment', ns: 'annul', type: 'annul:InvoiceAnnulment' },
  InvoiceData: { name: 'InvoiceData', ns: 'data', type: 'data:InvoiceData' },
  ManageAnnulmentRequest: {
    name: 'ManageAnnulmentRequest',
    ns: 'api',
    type: 'api:ManageAnnulmentRequest',
  },
  ManageAnnulmentResponse: {
    name: 'ManageAnnulmentResponse',
    ns: 'api',
    type: 'api:ManageAnnulmentResponse',
  },
  ManageInvoiceRequest: {
    name: 'ManageInvoiceRequest',
    ns: 'api',
    type: 'api:ManageInvoiceRequest',
  },
  ManageInvoiceResponse: {
    name: 'ManageInvoiceResponse',
    ns: 'api',
    type: 'api:ManageInvoiceResponse',
  },
  QueryInvoiceChainDigestRequest: {
    name: 'QueryInvoiceChainDigestRequest',
    ns: 'api',
    type: 'api:QueryInvoiceChainDigestRequest',
  },
  QueryInvoiceChainDigestResponse: {
    name: 'QueryInvoiceChainDigestResponse',
    ns: 'api',
    type: 'api:QueryInvoiceChainDigestResponse',
  },
  QueryInvoiceCheckRequest: {
    name: 'QueryInvoiceCheckRequest',
    ns: 'api',
    type: 'api:QueryInvoiceCheckRequest',
  },
  QueryInvoiceCheckResponse: {
    name: 'QueryInvoiceCheckResponse',
    ns: 'api',
    type: 'api:QueryInvoiceCheckResponse',
  },
  QueryInvoiceDataRequest: {
    name: 'QueryInvoiceDataRequest',
    ns: 'api',
    type: 'api:QueryInvoiceDataRequest',
  },
  QueryInvoiceDataResponse: {
    name: 'QueryInvoiceDataResponse',
    ns: 'api',
    type: 'api:QueryInvoiceDataResponse',
  },
  QueryInvoiceDigestRequest: {
    name: 'QueryInvoiceDigestRequest',
    ns: 'api',
    type: 'api:QueryInvoiceDigestRequest',
  },
  QueryInvoiceDigestResponse: {
    name: 'QueryInvoiceDigestResponse',
    ns: 'api',
    type: 'api:QueryInvoiceDigestResponse',
  },
  QueryServiceMetricsListResponse: {
    name: 'QueryServiceMetricsListResponse',
    ns: 'metrics',
    type: 'metrics:QueryServiceMetricsListResponse',
  },
  QueryServiceMetricsResponse: {
    name: 'QueryServiceMetricsResponse',
    ns: 'metrics',
    type: 'metrics:QueryServiceMetricsResponse',
  },
  QueryTaxpayerRequest: {
    name: 'QueryTaxpayerRequest',
    ns: 'api',
    type: 'api:QueryTaxpayerRequest',
  },
  QueryTaxpayerResponse: {
    name: 'QueryTaxpayerResponse',
    ns: 'api',
    type: 'api:QueryTaxpayerResponse',
  },
  QueryTransactionListRequest: {
    name: 'QueryTransactionListRequest',
    ns: 'api',
    type: 'api:QueryTransactionListRequest',
  },
  QueryTransactionListResponse: {
    name: 'QueryTransactionListResponse',
    ns: 'api',
    type: 'api:QueryTransactionListResponse',
  },
  QueryTransactionStatusRequest: {
    name: 'QueryTransactionStatusRequest',
    ns: 'api',
    type: 'api:QueryTransactionStatusRequest',
  },
  QueryTransactionStatusResponse: {
    name: 'QueryTransactionStatusResponse',
    ns: 'api',
    type: 'api:QueryTransactionStatusResponse',
  },
  TokenExchangeRequest: {
    name: 'TokenExchangeRequest',
    ns: 'api',
    type: 'api:TokenExchangeRequest',
  },
  TokenExchangeResponse: {
    name: 'TokenExchangeResponse',
    ns: 'api',
    type: 'api:TokenExchangeResponse',
  },
};